// src/settings/settings.interface.ts
export interface AngularRenamerSettings {
    updateImports: boolean;
    autoRenameClasses: boolean;
    renameComponents: boolean;
    renameServices: boolean;
    renameDirectives: boolean;
    renamePipes: boolean;
    renameSpec: boolean;
    searchAndReplaceDeeply:boolean
    renameGuards:boolean
}